+++ conftest.py
@@ -1,12 +1,7 @@
 import pytest
-
-from common.global_var import initGlobal
-from environment.test_env_impl import TestEnvironment
 from common import download
+from environment.env import create_env
 from common.log import log
-
-from conf.settings import PLATON_BIN_FILE
-
 
 
 @pytest.fixture(scope="module")
@@ -16,61 +11,33 @@ def consensus_test_env(global_test_env):
 
 
 def pytest_addoption(parser):
-    parser.addoption("--platon_url", action="store",  help="platon_url: url to download platon bin")
+    parser.addoption("--tmpDir", action="store", help="tmpDir: tmp dir, default in deploy/tmp/global")
+    parser.addoption("--platonUrl", action="store",  help="platonUrl: url to download platon bin")
     parser.addoption("--nodeFile", action="store",  help="nodeFile: the node config file")
     parser.addoption("--accountFile", action="store", help="accountFile: the accounts file")
-    parser.addoption("--initChain", action="store_true", default=False, dest="initChain", help="nodeConfig: default to init chain data")
-    parser.addoption("--startAll", action="store_true", default=False, dest="startAll", help="startAll: default to start all nodes")
+    parser.addoption("--initChain", action="store_true", default=True, dest="initChain", help="nodeConfig: default to init chain data")
     parser.addoption("--installDependency", action="store_true", default=False, dest="installDependency", help="installDependency: default do not install dependencies")
-    parser.addoption("--installSuperVisor", action="store_true", default=False, dest="installSuperVisor", help="installSuperVisor: default do not install supervisor service")
+    parser.addoption("--installSupervisor", action="store_true", default=False, dest="installSuperVisor", help="installSupervisor: default do not install supervisor service")
 
-# py.test test_start.py -s --concmode=asyncnet --nodeFile "deploy/4_node.yml" --accountFile "deploy/accounts.yml" --initChain --startAll
-@pytest.fixture(scope="session", autouse=True)
+# py.test test_start.py -s --concmode=asyncnet --nodeFile "deploy/4_node.yml" --accountFile "deploy/accounts.yml" --initChain
+# py.test 'tests/chain/test_chain_deploy.py' --nodeFile "deploy/node/test_chaininfo.yml" --accountFile "deploy/accounts.yml" --alluredir="report/allure" -s -v
+@pytest.fixture(scope="session", autouse=False)
 def global_test_env(request):
-
     log.info("global_test_env>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>")
-
-    nodeFile = request.config.getoption("--nodeFile")
-    accountFile = request.config.getoption("--accountFile")
-    initChain = request.config.getoption("--initChain")
-    startAll = request.config.getoption("--startAll")
-    installDependency = request.config.getoption("--installDependency")
-    installSuperVisor = request.config.getoption("--installSuperVisor")
-    plant_url = request.config.getoption("--platon_url")
+    tmp_dir = request.config.getoption("--tmpDir")
+    node_file = request.config.getoption("--nodeFile")
+    account_file = request.config.getoption("--accountFile")
+    init_chain = request.config.getoption("--initChain")
+    install_dependency = request.config.getoption("--installDependency")
+    install_supervisor = request.config.getoption("--installSupervisor")
+    plant_url = request.config.getoption("--platonUrl")
     if plant_url:
         download.download_platon(plant_url)
-    env = create_env_impl(PLATON_BIN_FILE, nodeFile,'global', accountFile, initChain, startAll, installDependency, installSuperVisor)
-
+    env = create_env(tmp_dir, node_file, account_file, init_chain, install_dependency, install_supervisor)
 
     yield env
 
-    #todo
-    #env.shutdown()
+    # todo
+    env.shutdown()
 
-'''
-@pytest.fixture(scope="function")
-def custom_test_env():
-    def _custom_test_env(conf):
-        _ = conf.get("binFile")
-        nodeFile = conf.get("nodeFile")
-        genesisFile = conf.get("genesisFile")
-        accountFile = conf.get("accountFile")
-        initChain = conf.get("initChain")
-        _ = conf.get("startAll")
-        _ = conf.get("isHttpRpc")
-        return create_env_impl(node_file=nodeFile, account_file=accountFile, init_chain=initChain)
-    yield _custom_test_env
-   # _custom_test_env.shutdown()
-'''
-
-
-
-def create_env_impl(binfile, nodeFile, confdir, accountFile,initChain=True, startAll=True, installDependency=False, installSuperVisor=False)->TestEnvironment:
-    env = TestEnvironment(node_file=nodeFile, bin_file=binfile, confdir=confdir, account_file=accountFile, init_chain=initChain, startAll=startAll,
-    install_dependency=installDependency, install_supervisor=installSuperVisor)
-    print(env.install_dependency)
-    print(env.install_supervisor)
-    env.deploy_all()
-    env.start_all()
-    return env
