+++ deploy/template/static-nodes.json
@@ -1,8 +1,8 @@
 [
-"enode://fcd131162655c17bf585a7b77e50012a70be77b7410e5983551a41dcb0b65bf9fc52299f51890eeaf517117e565fec0546418950848bcac29859948b206b0de7@192.168.9.201:16789",
-"enode://70741deaf0111a121fff46ed7becefc68e50e2f0d7de44d992a19a80161b1ef9583df6b4871013d4e3379e9ba3fb433e3916737ac646c3068789c3257953d583@192.168.9.202:16789",
-"enode://212aafe54595d10e05b45372191e7a8f570b6a93ad1f80d3c68fd9c180c075a5655ffdc83227a52f0b3cf9e43ff651d4913358b6dc6efca588c85862de85eea4@192.168.9.204:16789",
 "enode://69ff3e17e4011b5b0909fb98eaae958398a9e59592081002ee4333f4dcbab7d927892f63be5592bcb1f7c1e41f0fcd8fb0cdf1c2afac06b2da727830e66d2043@192.168.9.203:16789",
+"enode://212aafe54595d10e05b45372191e7a8f570b6a93ad1f80d3c68fd9c180c075a5655ffdc83227a52f0b3cf9e43ff651d4913358b6dc6efca588c85862de85eea4@192.168.9.204:16789",
+"enode://70741deaf0111a121fff46ed7becefc68e50e2f0d7de44d992a19a80161b1ef9583df6b4871013d4e3379e9ba3fb433e3916737ac646c3068789c3257953d583@192.168.9.202:16789",
+"enode://fcd131162655c17bf585a7b77e50012a70be77b7410e5983551a41dcb0b65bf9fc52299f51890eeaf517117e565fec0546418950848bcac29859948b206b0de7@192.168.9.201:16789",
 "enode://a5537f9556f5ab952a629c9b7bab0fae8b4693783869369fe79ada57642e9693822f300f56b37608f537027eb3177a42b4cb0fd7234f467221e8c0ee8118e179@192.168.9.201:16787",
 "enode://b8bc20315cb1bfa66309362906b01ad6f9356dd6c96b97082af45944cf0ef56ae07a2240549f5ab82ad5d08aa5019fc154645be9d2d6eeeb220f301128aa631b@192.168.112.205:16788"
 ]
