+++ tests/initialize/test_initialize.py
@@ -2,7 +2,7 @@ import os
 import pytest
 import json
 import allure
-from environment.test_env_impl import TestEnvironment
+from environment.env import create_env
 from conf.settings import NODE_FILE
 from common.log import log
 
@@ -10,9 +10,10 @@ BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 @pytest.fixture(scope="module")
 def global_env():
-    env = TestEnvironment(node_file=NODE_FILE)
+    env = create_env(node_file=NODE_FILE)
+    # env = TestEnvironment(node_file=NODE_FILE)
     env.deploy_all()
-    env.start_all()
+    # env.start_all()
     return env
